% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/im.R
\name{im}
\alias{im}
\alias{mine}
\title{Specify an intermine data warehouse}
\usage{
im(mine)
}
\arguments{
\item{mine}{character string; name of intermine data warehouse}
}
\value{
named list.
}
\description{
Specify an intermine data warehouse
}
\section{Supported Mines}{


Interminer currently supports a subset of InterMine's data warehouses:

\itemize{
\item "FlyMine"
\item "modMine"
\item "YeastMine"

}
}

