% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.R
\name{im_get}
\alias{im_get}
\alias{im_get_bed}
\alias{im_get_fasta}
\alias{im_get_gff3}
\title{GET query results}
\usage{
im_get(im, as = "tab", path = NULL)

im_get_gff3(im, path = NULL)

im_get_fasta(im, path = NULL)

im_get_bed(im, path = NULL)
}
\arguments{
\item{im}{im object from \code{\link{im}}}

\item{as}{character string of the format to accept}

\item{path}{character string (optional) specifying the path to write query
results to.}
}
\description{
GET query results
}
\section{as}{

If querying sequence features then one of "gff3", "fasta", "bed". For
arbitrary queries defaults to "tab".
}

