% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regions.R
\name{im_regions}
\alias{im_regions}
\title{add interval regions to im object}
\usage{
im_regions(im, name, features, organism, cse_cols = 1:3L)
}
\arguments{
\item{im}{im object from \code{\link{im}}}

\item{name}{character string path to interval file or name or R object}

\item{features}{character vector of intermine sequence feature objects}

\item{organism}{character string of organism to query}

\item{cse_cols}{integer vector specifying the index of chromosome, start and
end columns.}
}
\description{
add interval regions to im object
}

