% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse_bed_results}
\alias{parse_bed_results}
\title{parse bed results into data.frame}
\usage{
parse_bed_results(r)
}
\arguments{
\item{r}{request object}
}
\description{
parse bed results into data.frame
}
\keyword{internal}

