% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse_fasta_results}
\alias{parse_fasta_results}
\title{parse fasta results into named list}
\usage{
parse_fasta_results(r)
}
\arguments{
\item{r}{request object}
}
\description{
parse fasta results into named list
}
\keyword{internal}

