% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regions.R
\name{parse_features}
\alias{parse_features}
\title{Convert features to intermine acceptable format}
\usage{
parse_features(feature_types)
}
\arguments{
\item{feature_types}{character vector; each element must be an intermine
sequence feature.}
}
\value{
space-separated string of featureTypes.
}
\description{
Convert features to intermine acceptable format
}
\keyword{internal}

