% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare.R
\name{parse_for_service}
\alias{parse_for_service}
\title{parse im object in preparation for a given service}
\usage{
parse_for_service(im, as)
}
\arguments{
\item{im}{im object}
}
\description{
parse im object in preparation for a given service
}
\keyword{internal}

