% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse_gff3_results}
\alias{parse_gff3_results}
\title{parse gff3 results into data.frame}
\usage{
parse_gff3_results(r)
}
\arguments{
\item{r}{request object}
}
\description{
parse gff3 results into data.frame
}
\keyword{internal}

