% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regions.R
\name{parse_regions}
\alias{parse_regions}
\title{Convert regions to intermine acceptable format}
\usage{
parse_regions(name, cse_cols = 1:3L)
}
\arguments{
\item{name}{chracter string; path to .bed or .gff3 file or data.frame object
containing chromosome, start and end columns.}

\item{cse_cols}{integer vector; indicates the chromosome, start, end columns
of a data.frame or file.}
}
\value{
returns a character vector of with each region of the form "c:s..e".
}
\description{
Convert regions to intermine acceptable format
}
\keyword{internal}

