% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse_results}
\alias{parse_results}
\title{parse results of a request}
\usage{
parse_results(r, as, path)
}
\arguments{
\item{r}{request object}

\item{path}{character string (optional) specifying the path to write query
results to.}
}
\description{
parse results of a request
}
\keyword{internal}

