% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse_tab_results}
\alias{parse_tab_results}
\title{parse general tabular results into data.frame}
\usage{
parse_tab_results(r)
}
\arguments{
\item{r}{request object}
}
\description{
parse general tabular results into data.frame
}
\keyword{internal}

