% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{prep_query}
\alias{prep_query}
\title{prepare a query object}
\usage{
prep_query(view, model = "genomic", name = NULL, sortOrder = NULL,
  constraintLogic = NULL)
}
\arguments{
\item{view}{}

\item{model}{}
}
\description{
prepare a query object
}
\keyword{internal}

