% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare.R
\name{prep_service_list}
\alias{prep_service_list}
\title{parse im for list type query}
\usage{
prep_service_list(im, as)
}
\arguments{
\item{im}{im object}
}
\description{
parse im for list type query
}
\keyword{internal}

