% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/service.R
\name{prep_service_query}
\alias{prep_service_query}
\title{parse im for arbitrary query}
\usage{
prep_service_query(im, as)
}
\arguments{
\item{im}{im object}
}
\description{
parse im for arbitrary query
}
\keyword{internal}

