% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/service.R
\name{prep_service_regions}
\alias{prep_service_regions}
\title{parse im for regions type query}
\usage{
prep_service_regions(im, as)
}
\arguments{
\item{im}{im object}
}
\description{
parse im for regions type query
}
\keyword{internal}

