% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/im.R
\name{ws_url}
\alias{ws_url}
\title{url for an intermine data warehouse}
\usage{
ws_url(mine)
}
\arguments{
\item{mine}{character string; name of intermine data warehouse}
}
\value{
intermine url
}
\description{
url for an intermine data warehouse
}
\keyword{internal}

